requirements.txt
Flask==2.3.3
astroid==3.2.4
autopep8==2.0.4
blinker==1.8.2
certifi==2024.7.4
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
docstring-to-markdown==0.15
flake8==7.0.0
Flask==2.3.3
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
importlib_metadata==8.1.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
nco==1.1.2
netCDF4==1.7.1.post1
numpy==2.0.1
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
pylint==3.2.6
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
pytoolconfig==1.3.1
requests==2.32.3
rope==1.13.0
setuptools==69.0.3
smmap==5.0.1
snowballstemmer==2.2.0
tomli==2.0.1
tomlkit==0.13.0
ujson==5.10.0
urllib3==2.2.2
vercel_blob==0.2.5
Werkzeug==3.0.3
whatthepatch==1.0.6
wheel==0.43.0
yapf==0.40.2
zipp==3.19.2
waitress==1.2.1